Cluster park district hoops when exporting to geojson
Taking a closer look at the data, there's actually a record for
basketball courts and each backboard.  However, there appear to
be backboards independent of courts.  So, when dumping the data
to geojson, iterate through the backboard records and cluster them
so there aren't so many markers on top of each other.

Also, make the geojson export a method of the model's queryset instead
of being in the data module.