Load tract shapes into database
Add a fabric task and ORM mappings for loading and querying
Chicago Census Tract shapes from
https://data.cityofchicago.org/Facilities-Geographic-Boundaries/Boundaries-Census-Tracts-2010/biqm-wjk3

What I really want this for is to filter the values from the P12
census tables down to just tracts in Chicago (since the CSV data
loaded for the P12 table is for all of Illinois).

http://stackoverflow.com/a/15962062/386210 and
http://lucumr.pocoo.org/2011/7/19/sqlachemy-and-you/ were extremely
helpful in figuring out how to do this.